BUGFIX - EXPORTS - CruiseData dir confusing misnomer
This has caught me out before, CRUISEDATA_DIR_ON_DATA_WAREHOUSE variable was also used to point to local mount of cruisedata in filecrop section, it was not obvious that while most functions use ssh or rsync there also has to be an nfs mount of the same dir. Altogether there's three ways of accessing an nfs mounted dir.